tornado/day04/requirement.txt

PyMySQL==0.9.3
SQLAlchemy==1.3.1
tornado==4.5
baidu-aip
